requirements.txt
flask
groq
python-dotenv
opencv-python
pytesseract
pillow
sympy
numpy